+++ packages/runtime/src/runtime/hooks/useEnsembleAction.tsx
@@ -3,25 +3,25 @@ import {
   useScreenContext,
   evaluate,
   isExpression,
+  useRegisterBindings,
 } from "framework";
 import type {
   InvokeAPIAction,
   ExecuteCodeAction,
   Response,
   EnsembleAction,
+  PickFilesAction,
+  UploadFilesAction,
+  ScreenContextDefinition,
 } from "framework";
-import { isString, merge } from "lodash-es";
-import { useState, useEffect, useMemo } from "react";
+import { head, isString, merge } from "lodash-es";
+import { useState, useEffect, useMemo, useCallback } from "react";
 import { useNavigateScreen } from "./useNavigateScreen";
 // FIXME: refactor
 // eslint-disable-next-line import/no-cycle
 import { useNavigateModalScreen } from "./useNavigateModal";
 import { useShowToast } from "./useShowToast";
 import { useCloseAllDialogs } from "./useCloseAllDialogs";
-// eslint-disable-next-line import/no-cycle
-import { usePickFiles } from "./usePickFiles";
-// eslint-disable-next-line import/no-cycle
-import { useUploadFiles } from "./useUploadFiles";
 
 export type EnsembleActionHookResult =
   | {
@@ -33,6 +33,13 @@ export type EnsembleActionHook<
   Q = unknown,
   R extends EnsembleActionHookResult = EnsembleActionHookResult,
 > = (action?: T, options?: Q) => R;
+
+type UploadStatus =
+  | "pending"
+  | "running"
+  | "completed"
+  | "cancelled"
+  | "failed";
 
 export interface UseExecuteCodeActionOptions {
   context?: Record<string, unknown>;
@@ -122,6 +129,163 @@ export const useInvokeApi: EnsembleActionHook<InvokeAPIAction> = (action) => {
   }, [action?.onError, error, onErrorAction]);
 
   return invokeApi;
+};
+
+export const usePickFiles: EnsembleActionHook<PickFilesAction> = (
+  action?: PickFilesAction,
+) => {
+  const [files, setFiles] = useState<FileList>();
+  const onCompleteAction = useEnsembleAction(action?.onComplete);
+
+  useRegisterBindings(
+    {
+      files,
+    },
+    action?.id,
+    {
+      setFiles,
+    },
+  );
+
+  const inputEl = useMemo(() => {
+    const input = document.createElement("input");
+    input.type = "file";
+    input.multiple = action?.allowMultiple || false;
+    input.accept =
+      action?.allowedExtensions?.map((ext) => ".".concat(ext))?.toString() ||
+      "*/*";
+
+    input.onchange = (event: Event): void => {
+      const selectedFiles =
+        (event.target as HTMLInputElement).files || undefined;
+
+      if (selectedFiles) setFiles(selectedFiles);
+
+      if (onCompleteAction) onCompleteAction?.callback();
+    };
+
+    return input;
+  }, [action?.allowMultiple, action?.allowedExtensions, onCompleteAction]);
+
+  const callback = useCallback((): void => {
+    try {
+      inputEl.click();
+    } catch (error) {
+      // eslint-disable-next-line no-console
+      console.error(error);
+    }
+  }, [inputEl]);
+
+  return { callback };
+};
+
+export const useUploadFiles: EnsembleActionHook<UploadFilesAction> = (
+  action?: UploadFilesAction,
+) => {
+  const [body, setBody] = useState<Record<string, unknown>>();
+  const [headers, setHeaders] = useState<Record<string, unknown>>();
+  const [status, setStatus] = useState<UploadStatus>("pending");
+  const [progress, setProgress] = useState<number>(0.0);
+
+  const onCompleteAction = useEnsembleAction(action?.onComplete);
+  const onErrorAction = useEnsembleAction(action?.onError);
+  const screenContext = useScreenContext();
+
+  useRegisterBindings(
+    {
+      body,
+      headers,
+      status,
+      progress,
+    },
+    action?.id,
+    {
+      setBody,
+      setHeaders,
+      setStatus,
+      setProgress,
+    },
+  );
+
+  const apiModel = screenContext?.model?.apis?.find(
+    (model) => model?.name === action?.uploadApi,
+  );
+  if (!apiModel) return;
+
+  const progressCallback = (progressEvent: ProgressEvent): void => {
+    const percentCompleted = (progressEvent.loaded * 100) / progressEvent.total;
+
+    setProgress(percentCompleted);
+  };
+
+  const callback = async (): Promise<void> => {
+    const files = evaluate(
+      screenContext as ScreenContextDefinition,
+      action?.files,
+    ) as FileList | undefined;
+    if (!files || files.length === 0) throw Error("Files not found");
+
+    const formData = new FormData();
+    if (files.length === 1)
+      formData.append(action?.fieldName ?? "files", head(files) as Blob);
+    else
+      for (let i = 0; i < files.length; i++) {
+        formData.append(action?.fieldName ?? `file${i}`, files[i]);
+      }
+
+    const apiModelBody = apiModel?.body ?? {};
+    for (const key in apiModelBody) {
+      formData.append(key, apiModelBody[key] as string);
+    }
+    const resolvedInputs = Object.entries(action?.inputs ?? {})
+      .concat(
+        (apiModel?.inputs?.map((value) => [value, value]) as [
+          string,
+          unknown,
+        ][]) ?? [],
+      )
+      .map(([key, value]) => {
+        if (isExpression(value)) {
+          const resolvedValue = evaluate(
+            screenContext as ScreenContextDefinition,
+            value,
+          );
+          return [key, resolvedValue];
+        }
+        return [key, value];
+      });
+
+    try {
+      setStatus("running");
+      const response = await DataFetcher.uploadFiles(
+        apiModel.uri,
+        apiModel.method,
+        {
+          "Content-Type": "multipart/form-data",
+          ...apiModel.headers,
+        },
+        formData,
+        Object.fromEntries(resolvedInputs) as Record<string, unknown>,
+        progressCallback,
+      );
+
+      setBody(response.body as Record<string, unknown>);
+      setHeaders(response.headers as Record<string, unknown>);
+      if (response.isSuccess) {
+        setStatus("completed");
+        onCompleteAction?.callback();
+      } else {
+        setStatus("failed");
+        onErrorAction?.callback();
+      }
+    } catch (err: unknown) {
+      setBody(err as Record<string, unknown>);
+      setStatus("failed");
+      onErrorAction?.callback();
+    }
+  };
+
+  return { callback };
 };
 
 /* eslint-disable react-hooks/rules-of-hooks */
